+++ stories/overview/cc-tile-status-codes.stories.js
@@ -1,4 +1,5 @@
 import '../../src/overview/cc-tile-status-codes.js';
+import '../../src/overview/cc-tile-status-codes.smart.js';
 import { makeStory, storyWait } from '../lib/make-story.js';
 import { enhanceStoriesNames } from '../lib/story-names.js';